+++ Source/PhatSDK/LandDefs.cpp
@@ -1,45 +1,45 @@
 
-#include "StdAfx.h"
+#include <StdAfx.h>
 #include "LandDefs.h"
 
 namespace LandDefs
 {
 	// Initialized from beginning.
-	DWORD blockid_mask = 0xFFFF0000;
-	DWORD lbi_cell_id = 0x0000FFFE;
-	DWORD block_cell_id = 0x0000FFFF;
-	DWORD first_envcell_id = 0x100;
-	DWORD last_envcell_id = 0xFFFD;
-	DWORD first_lcell_id = 1;
-	DWORD last_lcell_id = 64;
-	long max_block_width = 0xFF;
-	DWORD max_block_shift = 8;
-	DWORD blockx_mask = 0xFF00;
-	DWORD blocky_mask = 0x00FF;
-	DWORD block_part_shift = 16;
-	DWORD cellid_mask = 0x0000FFFF;
-	DWORD terrain_byte_offset = 2;
+	uint32_t blockid_mask = 0xFFFF0000;
+	uint32_t lbi_cell_id = 0x0000FFFE;
+	uint32_t block_cell_id = 0x0000FFFF;
+	uint32_t first_envcell_id = 0x100;
+	uint32_t last_envcell_id = 0xFFFD;
+	uint32_t first_lcell_id = 1;
+	uint32_t last_lcell_id = 64;
+	int32_t max_block_width = 0xFF;
+	uint32_t max_block_shift = 8;
+	uint32_t blockx_mask = 0xFF00;
+	uint32_t blocky_mask = 0x00FF;
+	uint32_t block_part_shift = 16;
+	uint32_t cellid_mask = 0x0000FFFF;
+	uint32_t terrain_byte_offset = 2;
 
 	// Initialized later on.
-	long side_vertex_count = 9;
+	int32_t side_vertex_count = 9;
 	float half_square_length = 12.0f;
 	float square_length = 24.0f;
 	float block_length = 192.0f;
-	long lblock_shift = 3;
-	long lblock_side = 8;
-	long lblock_mask = 7;
-	long land_width = 0x7F8;
-	long land_length = 0x7F8;
-	long num_block_length = 0xFF;
-	long num_block_width = 0xFF;
-	long num_blocks = 0xFF * 0xFF;
+	int32_t lblock_shift = 3;
+	int32_t lblock_side = 8;
+	int32_t lblock_mask = 7;
+	int32_t land_width = 0x7F8;
+	int32_t land_length = 0x7F8;
+	int32_t num_block_length = 0xFF;
+	int32_t num_block_width = 0xFF;
+	int32_t num_blocks = 0xFF * 0xFF;
 	float inside_val; // ? ....
 	float outside_val;
 	float max_object_height;
 	float road_width;
 	float sky_height;
-	long vertex_per_cell = 1;
-	long polys_per_landcell = 2;
+	int32_t vertex_per_cell = 1;
+	int32_t polys_per_landcell = 2;
 	float Land_Height_Table[256];
 
 	class FillHeightTable
@@ -56,13 +56,13 @@ namespace LandDefs
 			Land_Height_Table[i] = (float)(i * 2);
 	}
 
-	Vector get_block_offset(DWORD LandBlock1, DWORD LandBlock2)
+	Vector get_block_offset(uint32_t LandBlock1, uint32_t LandBlock2)
 	{
 		if ((LandBlock1 >> block_part_shift) == (LandBlock2 >> block_part_shift))
 			return Vector(0, 0, 0);
 
-		long X1, Y1;
-		long X2, Y2;
+		int32_t X1, Y1;
+		int32_t X2, Y2;
 		blockid_to_lcoord(LandBlock1, X1, Y1);
 		blockid_to_lcoord(LandBlock2, X2, Y2);
 
@@ -72,7 +72,7 @@ namespace LandDefs
 			0.0);
 	}
 
-	BOOL blockid_to_lcoord(DWORD LandBlock, long& X, long& Y)
+	BOOL blockid_to_lcoord(uint32_t LandBlock, int32_t& X, int32_t& Y)
 	{
 		if (!LandBlock)
 		{
@@ -93,7 +93,7 @@ namespace LandDefs
 		return TRUE;
 	}
 
-	BOOL gid_to_lcoord(DWORD LandCell, long& X, long& Y)
+	BOOL gid_to_lcoord(uint32_t LandCell, int32_t& X, int32_t& Y)
 	{
 		if (!inbound_valid_cellid(LandCell))
 			return FALSE;
@@ -113,15 +113,15 @@ namespace LandDefs
 		return TRUE;
 	}
 
-	BOOL inbound_valid_cellid(DWORD LandCell)
+	BOOL inbound_valid_cellid(uint32_t LandCell)
 	{
-		DWORD CellID = LandCell & cellid_mask;
+		uint32_t CellID = LandCell & cellid_mask;
 
 		if (((CellID >= first_lcell_id) && (CellID <= last_lcell_id)) ||
 			((CellID >= first_envcell_id) && (CellID <= last_envcell_id)) ||
 			((CellID == block_cell_id)))
 		{
-			long blockx;
+			int32_t blockx;
 
 			blockx = ((LandCell >> block_part_shift) & blockx_mask) >> max_block_shift;
 			blockx = blockx << lblock_shift;
@@ -135,18 +135,18 @@ namespace LandDefs
 		return FALSE;
 	}
 
-	DWORD lcoord_to_gid(long X, long Y)
+	uint32_t lcoord_to_gid(int32_t X, int32_t Y)
 	{
 		if (X < 0 || Y < 0 || X >= land_width || Y >= land_length)
 			return 0;
 
-		DWORD Block = ((X >> lblock_shift) << max_block_shift) | (Y >> lblock_shift);
-		DWORD Cell = first_lcell_id + ((X & lblock_mask) << lblock_shift) + (Y & lblock_mask);
+		uint32_t Block = ((X >> lblock_shift) << max_block_shift) | (Y >> lblock_shift);
+		uint32_t Cell = first_lcell_id + ((X & lblock_mask) << lblock_shift) + (Y & lblock_mask);
 
-		return (DWORD)((Block << block_part_shift) | Cell);
+		return (uint32_t)((Block << block_part_shift) | Cell);
 	}
 
-	BOOL in_bounds(long GidX, long GidY)
+	BOOL in_bounds(int32_t GidX, int32_t GidY)
 	{
 		if (GidX < 0 || GidY < 0)
 			return FALSE;
@@ -156,16 +156,16 @@ namespace LandDefs
 		return TRUE;
 	}
 
-	DWORD get_block_gid(long X, long Y)
+	uint32_t get_block_gid(int32_t X, int32_t Y)
 	{
 		if (!in_bounds(X, Y))
 			return 0;
 
-		return ((DWORD)(((X >> lblock_shift) << max_block_shift) | (Y >> lblock_shift)) << block_part_shift) | block_cell_id;
+		return ((uint32_t)(((X >> lblock_shift) << max_block_shift) | (Y >> lblock_shift)) << block_part_shift) | block_cell_id;
 	}
 
-	BOOL init(long NumBlockLength, long NumBlockWidth, float SquareLength, long LBlockSide,
-		long VertexPerCell, float MaxObjHeight, float SkyHeight, float RoadWidth)
+	BOOL init(int32_t NumBlockLength, int32_t NumBlockWidth, float SquareLength, int32_t LBlockSide,
+		int32_t VertexPerCell, float MaxObjHeight, float SkyHeight, float RoadWidth)
 	{
 		if (NumBlockLength <= 0 || NumBlockLength > max_block_width)
 			return FALSE;
@@ -245,9 +245,9 @@ namespace LandDefs
 		return TRUE;
 	}
 
-	BOOL get_outside_lcoord(DWORD cellid, Vector *_offset, long *x, long *y)
+	BOOL get_outside_lcoord(uint32_t cellid, Vector *_offset, int32_t *x, int32_t *y)
 	{
-		DWORD cell = cellid & cellid_mask;
+		uint32_t cell = cellid & cellid_mask;
 
 		if ((cell >= first_lcell_id && cell <= last_lcell_id) ||
 			(cell >= first_envcell_id && cell <= last_envcell_id) ||
@@ -255,8 +255,8 @@ namespace LandDefs
 		{
 			blockid_to_lcoord(cellid, *x, *y);
 
-			*x = *x + (long)floor(_offset->x / square_length);
-			*y = *y + (long)floor(_offset->y / square_length);
+			*x = *x + (int32_t)floor(_offset->x / square_length);
+			*y = *y + (int32_t)floor(_offset->y / square_length);
 
 			if (*x < 0 || *y < 0 || *x >= land_width || *y >= land_length)
 				return FALSE;
@@ -267,11 +267,11 @@ namespace LandDefs
 			return FALSE;
 	}
 
-	BOOL adjust_to_outside(DWORD *_cellid, Vector *_offset)
+	BOOL adjust_to_outside(uint32_t *_cellid, Vector *_offset)
 	{
-		DWORD cellid = *_cellid;
+		uint32_t cellid = *_cellid;
 
-		DWORD cell = cellid & cellid_mask;
+		uint32_t cell = cellid & cellid_mask;
 
 		if ((cell >= first_lcell_id && cell <= last_lcell_id) ||
 			(cell >= first_envcell_id && cell <= last_envcell_id) ||
@@ -282,7 +282,7 @@ namespace LandDefs
 			if (fabs(_offset->y) < F_EPSILON)
 				_offset->y = 0;
 
-			long x, y;
+			int32_t x, y;
 			if (get_outside_lcoord(*_cellid, _offset, &x, &y))
 			{
 				*_cellid = lcoord_to_gid(x, y);
